+++ NEWS.md
@@ -1,12 +1,20 @@
-## 3.0.0 (2019-08-08)
+## 3.0.0 (2019-08-15)
 
   Competition Doom 3.0 is based on Chocolate Doom 3.1.
   See Chocolate Doom 3.0 history for changes:
   https://github.com/chocolate-doom/chocolate-doom
+
   Imported most of v2 Competition features into v3:
   - naming, icons & settings
   - in game & end screen timer, novert always on
+  - show stats screen after EXM8
+  - enable recording Doom movie
+  - using -printstats & -record outputs stats to stats.txt
+
   IN PROGRESS/TODO:
-  - recording episode
   - demo metadata
   - competition setup menu
+  - high resolution
+
+  BUGS:
+ - if demo filename is longer than 256 chars it won't be not recorded